Simplify authorization checks
