+++ app/src/main/java/com/sama/common/Collections.kt
@@ -1,19 +1,10 @@
 package com.sama.common
 
-import java.time.Duration
-import java.time.Instant
-import java.time.LocalDate
-import java.time.ZoneId
 import java.util.Objects
 import java.util.Optional
-import kotlin.streams.asSequence
-import org.springframework.data.repository.CrudRepository
-import org.threeten.extra.LocalDateRange
 
 fun <T : Any> Optional<T>.toNullable(): T? = this.orElse(null)
 
-inline fun <reified T, ID> CrudRepository<T, ID>.findByIdOrThrow(id: ID): T = findById(id)
-    .orElseThrow { NotFoundException(T::class, id) }
 
 fun <T> MutableList<T>.mapMutating(transform: (T) -> T): MutableList<T> {
     Objects.requireNonNull(transform)
@@ -90,34 +81,4 @@ fun <T> chunkedByIterator(
             yield(buffer)
         }
     }
-}
-
-
-fun Long.toMinutes(): Duration = Duration.ofMinutes(this)
-
-fun LocalDate.datesUtil(endDate: LocalDate): Sequence<LocalDate> {
-    return this.datesUntil(endDate).asSequence()
-}
-
-operator fun LocalDateRange.component1(): LocalDate = this.start
-operator fun LocalDateRange.component2(): LocalDate = this.end
-
-@JvmName("to")
-infix fun LocalDate.to(that: LocalDate): LocalDateRange = LocalDateRange.of(this, that)
-
-@JvmName("toNullable")
-infix fun LocalDate?.to(that: LocalDate?): LocalDateRange? {
-    return if (this == null || that == null) null
-    else LocalDateRange.of(this, that)
-}
-
-fun ZoneId.toGmtString(atDate: Instant): String {
-    val offsetSeconds = this.rules.getOffset(atDate).totalSeconds
-    val plusSign = if (offsetSeconds > 0) "+" else ""
-    val offsetString = when {
-        offsetSeconds == 0 -> ""
-        offsetSeconds.mod(3600) == 0 -> "${offsetSeconds / 3600}"
-        else -> "${offsetSeconds.toFloat() / 3600}"
-    }
-    return "GMT$plusSign$offsetString"
 }
